backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.1.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow-intel
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   starlette
    #   watchfiles
astunparse==1.6.3
    # via tensorflow-intel
beautifulsoup4==4.13.3
    # via gdown
blinker==1.9.0
    # via flask
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   flask
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   colorlog
    #   tqdm
    #   uvicorn
colorlog==6.9.0
    # via backend (pyproject.toml)
deepface==0.0.93
    # via backend (pyproject.toml)
fastapi==0.115.8
    # via backend (pyproject.toml)
filelock==3.17.0
    # via gdown
fire==0.7.0
    # via deepface
flask==3.1.0
    # via
    #   deepface
    #   flask-cors
flask-cors==5.0.0
    # via deepface
flatbuffers==25.1.24
    # via tensorflow-intel
gast==0.6.0
    # via tensorflow-intel
gdown==5.2.0
    # via
    #   deepface
    #   retina-face
google-pasta==0.2.0
    # via tensorflow-intel
greenlet==3.1.1
    # via sqlalchemy
grpcio==1.70.0
    # via
    #   tensorboard
    #   tensorflow-intel
gunicorn==23.0.0
    # via deepface
h11==0.14.0
    # via uvicorn
h5py==3.12.1
    # via
    #   keras
    #   tensorflow-intel
httptools==0.6.4
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
joblib==1.4.2
    # via mtcnn
keras==3.8.0
    # via
    #   deepface
    #   tensorflow-intel
libclang==18.1.1
    # via tensorflow-intel
lz4==4.4.3
    # via mtcnn
markdown==3.7
    # via tensorboard
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.4.1
    # via
    #   keras
    #   tensorflow-intel
mtcnn==1.0.0
    # via deepface
namex==0.0.8
    # via keras
numpy==2.0.2
    # via
    #   deepface
    #   h5py
    #   keras
    #   ml-dtypes
    #   opencv-python
    #   pandas
    #   retina-face
    #   tensorboard
    #   tensorflow-intel
opencv-python==4.11.0.86
    # via
    #   deepface
    #   retina-face
opt-einsum==3.4.0
    # via tensorflow-intel
optree==0.14.0
    # via keras
packaging==24.2
    # via
    #   gunicorn
    #   keras
    #   tensorboard
    #   tensorflow-intel
pandas==2.2.3
    # via deepface
pillow==11.1.0
    # via
    #   backend (pyproject.toml)
    #   deepface
    #   retina-face
protobuf==5.29.3
    # via
    #   tensorboard
    #   tensorflow-intel
pydantic==2.10.6
    # via
    #   fastapi
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pysocks==1.7.1
    # via requests
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via
    #   backend (pyproject.toml)
    #   uvicorn
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via uvicorn
requests==2.32.3
    # via
    #   deepface
    #   gdown
    #   tensorflow-intel
retina-face==0.0.17
    # via deepface
rich==13.9.4
    # via keras
setuptools==75.8.0
    # via
    #   tensorboard
    #   tensorflow-intel
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorboard
    #   tensorflow-intel
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.37
    # via sqlmodel
sqlmodel==0.0.22
    # via backend (pyproject.toml)
starlette==0.45.3
    # via fastapi
tensorboard==2.18.0
    # via tensorflow-intel
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.18.0
    # via
    #   backend (pyproject.toml)
    #   deepface
    #   retina-face
    #   tf-keras
tensorflow-intel==2.18.0
    # via
    #   backend (pyproject.toml)
    #   tensorflow
termcolor==2.5.0
    # via
    #   fire
    #   tensorflow-intel
tf-keras==2.18.0
    # via backend (pyproject.toml)
tqdm==4.67.1
    # via
    #   deepface
    #   gdown
typing-extensions==4.12.2
    # via
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   optree
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   tensorflow-intel
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via backend (pyproject.toml)
watchfiles==1.0.4
    # via uvicorn
websockets==14.2
    # via uvicorn
werkzeug==3.1.3
    # via
    #   flask
    #   tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.2
    # via tensorflow-intel
